update install scripts for windows
