+++ binding.gyp
@@ -89,7 +89,7 @@
         'OS=="win"', {
           'defines': ['COMPILER_MSVC'],
           'libraries': ['tensorflow'],
-          'library_dirs' : ['<(INTERMEDIATE_DIR)'],
+          'library_dirs' : ['<(module_path)'],
           'variables': {
             'tensorflow-library-target': 'windows'
           },
@@ -106,13 +106,13 @@
                 '<(module_root_dir)/scripts/deps-stage.js'
               ],
               'outputs': [
-                '<(PRODUCT_DIR)/tensorflow.dll',
+                '<(module_path)/tensorflow.dll',
               ],
               'action': [
                 'node',
                 '<@(_inputs)',
                 '<@(tensorflow-library-action)',
-                '<(PRODUCT_DIR)'
+                '<(module_path)'
               ]
             },
             {
@@ -120,10 +120,10 @@
               'inputs': [
                 '<(module_root_dir)/scripts/generate_defs.js',
                 '<@(tensorflow_headers)',
-                "<(PRODUCT_DIR)/tensorflow.dll"
+                "<(module_path)/tensorflow.dll"
               ],
               'outputs': [
-                '<(INTERMEDIATE_DIR)/tensorflow.def'
+                '<(module_path)/tensorflow.def'
               ],
               'action': [
                 'cmd',
@@ -133,10 +133,10 @@
             {
               'action_name': 'build-tensorflow-lib',
               'inputs': [
-                '<(INTERMEDIATE_DIR)/tensorflow.def'
+                '<(module_path)/tensorflow.def'
               ],
               'outputs': [
-                '<(INTERMEDIATE_DIR)/tensorflow.lib'
+                '<(module_path)/tensorflow.lib'
               ],
               'action': [
                 'lib',
