+++ package.json
@@ -16,10 +16,10 @@
     "build-npm": "./scripts/build-npm.sh",
     "clean-deps": "rm -rf deps",
     "coverage": "nyc ts-node src/run_tests.ts",
-    "enable-gpu": "node scripts/install.js gpu download && yarn",
+    "enable-gpu": "node scripts/install-from-source.js gpu download && yarn",
     "format": "clang-format -i -style=Google binding/*.cc binding/*.h",
-    "install-local": "node scripts/install.js",
-    "install": "./scripts/user-install.sh",
+    "install-from-source": "node scripts/install-from-source.js",
+    "install": "node scripts/install.js",
     "link-local": "yalc link",
     "lint": "tslint -p . -t verbose",
     "prep": "cd node_modules/@tensorflow/tfjs-core && yarn && yarn build",
@@ -50,6 +50,7 @@
     "@tensorflow/tfjs": "~1.1.2",
     "adm-zip": "^0.4.11",
     "bindings": "~1.3.0",
+    "edit-json-file": "^1.2.1",
     "https-proxy-agent": "^2.2.1",
     "progress": "^2.0.0",
     "rimraf": "^2.6.2",
@@ -61,9 +62,9 @@
     "module_path": "./lib/binding/napi-v{napi_build_version}",
     "host": "https://storage.googleapis.com/tf-builds/pre-built-binary",
     "remote_path": "./napi-v{napi_build_version}/{version}/",
-    "package_name": "temp_package_name",
     "napi_versions": [
       3
-    ]
+    ],
+    "package_name": "CPU-windows-1.1.2.zip"
   }
 }
